Refactored session
